listo para hacer merge con main

+++ src/posts/posts.controller.ts
@@ -3,11 +3,8 @@ import {
   Get,
   Post,
   Body,
-  Patch,
   Param,
   Delete,
-  HttpException,
-  HttpStatus,
   Put,
   UseGuards,
   ForbiddenException,
@@ -37,6 +34,9 @@ export class PostsController {
   // búsqueda de posts, lo puse primero porque sino daba ----
   // error la url con la de GET por :id -----
   @Get('search')
+  @ApiOperation({
+    summary: 'Buscar Posts por título, contenido y más. Permite paginación.',
+  })
   searchPosts(
     @Query('query') query: string,
     @Query('page', new DefaultValuePipe(1), ParseIntPipe) page: number,
@@ -45,8 +45,11 @@ export class PostsController {
     return this.postsService.searchPosts(query, { page, limit });
   }
 
-
   @Get('filter')
+  @ApiOperation({
+    summary:
+      'Filtra posts por autor o categoría (type) y un texto dado (query)',
+  })
   async filterPosts(
     @Query('type') type: string,
     @Query('query') query: string,
@@ -60,7 +63,6 @@ export class PostsController {
     }
   }
 
-
   @Post()
   @ApiOperation({ summary: 'Crear un nuevo post' })
   @ApiBody({ type: CreatePostDto, description: 'Datos del post a crear' })
@@ -69,16 +71,26 @@ export class PostsController {
   }
 
   @Get()
+  @ApiOperation({
+    summary: 'Devuelve todos los posts',
+  })
   findAll(): Promise<PostInterface[]> | null {
     return this.postsService.findAll();
   }
 
   @Get(':id')
+  @ApiOperation({
+    summary: 'Devuelve un post específico ne base a su identificador',
+  })
   findOne(@Param('id') id: string): Promise<PostInterface> | null {
     return this.postsService.findOne(id);
   }
 
   @Put(':id')
+  @ApiOperation({
+    summary:
+      'Actualiza los datos de un post específico. Solo para usuarios ADMIN o el creador del post a actualizar.',
+  })
   @UseGuards(JwtAuthGuard, AdminAuthGuard)
   async update(
     @Param('id') id: string,
@@ -104,12 +116,35 @@ export class PostsController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string): Promise<PostInterface> | null {
+  @ApiOperation({
+    summary: 'Elimina un post. Solo para usuarios ADMIN o el creador del post a actualizar.',
+  })
+  async remove(
+    @Param('id') id: string,
+    @Req() req,
+  ): Promise<PostInterface> | null {
+    // Obtener el usuario desde el token JWT -----
+    const user = req.user;
+    const post = await this.postsService.findOne(id);
+
+    // Verificar si el usuario es un administrador o el creador del post
+    const isAdmin = user.roles.includes('ADMIN');
+    const isPostCreator = post.user_id === user.id;
+
+    if (!isAdmin && !isPostCreator) {
+      throw new ForbiddenException(
+        'No tienes permisos para realizar esta acción',
+      );
+    }
+
     return this.postsService.remove(id);
   }
 
   // obtener por userid -----
   @Get('user/:userId')
+  @ApiOperation({
+    summary: 'Devuelve todos los posts de un usuario dado.',
+  })
   async getPostsByUser(@Param('userId') userId: string) {
     const posts = await this.postsService.getPostsByUser(userId);
     return posts;
